Add ZSink.fromFile functionality (unified)
